requirements.txt
aiohttp
furl
loguru
PyYAML
qrcode
requests
rich
pyppeteer
pillow
moment
python-telegram-bot
opencv-python-headless
ujson
peewee
pyexecjs